Large-scale refactor; add package setup; debug bytes/str errors
