+++ appveyor.yml
@@ -82,13 +82,15 @@ install:
   - "ECHO %PYTHON% %PYTHON_VERSION%%APPVEYOR_BUILD_FOLDER%"
   - "set HOME=%APPVEYOR_BUILD_FOLDER%"
   - "%PYPY%"
-  - "%PIP% install codecov rsa"
+  - "%PIP% install pytest-coverage codecov rsa"
   - "cd %HOME%"
 
 test: off
 
 build_script:
-  - "%COV% run -a py/test.py"
+  - "%PYTHON% setup.py build"
+  - "cd build\\lib"
+  - "%PYTHON% -m pytest -vv --cov=./p2p/"
 
 on_success:
   - "%COV% combine"
