+++ py/__init__.py
@@ -1,0 +1,7 @@
+from .mesh import mesh_socket
+# from .chord import chord_socket
+from .net import uses_RSA, decryption_error, verification_error, newkeys,\
+                 encrypt, decrypt, sign, verify, secure_socket
+from .base import version as __version__
+
+__all__ = ["mesh", "chord", "net", "base"]
