+++ py/test/__init__.py
@@ -1,0 +1,3 @@
+from . import test_base, test_net
+
+__all__ = ['test_base', 'test_net']
