+++ setup.py
@@ -1,0 +1,25 @@
+from distutils.core import setup
+from py import __version__
+
+setup(name='p2p',
+      description='A python library for peer-to-peer networking',
+      version=__version__,
+      author='Gabe Appleton',
+      author_email='gappleto97+development@gmail.com',
+      url='https://github.com/gappleto97/p2p-project',
+      license='LGPL',
+      packages=['p2p', 'p2p.test'],
+      package_dir={'p2p': 'py'},
+      classifiers=['Development Status :: 2 - Pre-Alpha',
+                   'Intended Audience :: Developers',
+                   'License :: OSI Approved :: GNU Lesser General Public License v3 (LGPLv3)',
+                   'Programming Language :: Python :: 2',
+                   'Programming Language :: Python :: 2.6',
+                   'Programming Language :: Python :: 2.7',
+                   'Programming Language :: Python :: 3',
+                   'Programming Language :: Python :: 3.3',
+                   'Programming Language :: Python :: 3.4',
+                   'Programming Language :: Python :: 3.5',
+                   'Operating System :: OS Independent',
+                   'Programming Language :: Python',
+                   'Programming Language :: JavaScript'])
