+++ shippable.yml
@@ -16,10 +16,11 @@ matrix:
 
 build:
   ci:
-    - pip install codecov
+    - pip install pytest-coverage codecov
     - if [ $(($RANDOM % 2)) == 0 ] || [ $SHIPPABLE_PYTHON_VERSION == pypy ] || [ $SHIPPABLE_PYTHON_VERSION == pypy3 ]; then pip install rsa; else pip install PyCrypto; fi
-    - coverage run -a py/test.py
+    - python setup.py build
+    - cd build/lib
+    - py.test -vv --cov=./p2p/
   on_success:
     - coverage combine
     - codecov --token=d89f9bd9-27a3-4560-8dbb-39ee3ba020a5
-
